requirements.txt
# FastAPI stack
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.6,<1.0.0

# OpenAI
openai>=1.12.0,<2.0.0

# PDF & text
pdfplumber>=0.10.0,<1.0.0

# Embeddings tokenization
tiktoken>=0.5.2,<1.0.0

# Vector search
faiss-cpu>=1.7.4,<2.0.0
numpy>=1.26.0,<3.0.0

# Config
pydantic-settings>=2.1.0,<3.0.0
